+++ glue/clients/tests/test_histogram_client.py
@@ -14,6 +14,7 @@ from ...core.data_collection import DataCollection
 from ...core.exceptions import IncompatibleDataException
 from ...core.data import Data, CategoricalComponent, ComponentID
 from ...core.subset import RangeSubsetState, CategoricalRoiSubsetState
+from ...core.roi import RangeROI, PolygonalROI
 
 from .util import renderless_figure
 
@@ -222,9 +223,7 @@ class TestHistogramClient(object):
         # bins are -7...-1
 
         self.data.edit_subset = [self.data.subsets[0]]
-
-        roi = MagicMock()
-        roi.to_polygon.return_value = [-5.1, -4.5, -3.2], [2, 3, 4]
+        roi = PolygonalROI(vx=[-5.1, -4.5, -3.2], vy=[2, 3, 4])
 
         self.client.apply_roi(roi)
         state = self.data.subsets[0].subset_state
@@ -239,8 +238,7 @@ class TestHistogramClient(object):
         self.client.set_component(self.data.id['x'])
         self.data.edit_subset = [self.data.subsets[0]]
         self.client.xlog = True
-        roi = MagicMock()
-        roi.to_polygon.return_value = [1, 2, 3], [2, 3, 4]
+        roi = PolygonalROI(vx=[1, 2, 3], vy=[2, 3, 4])
 
         self.client.apply_roi(roi)
         state = self.data.subsets[0].subset_state