+++ glue/core/data.py
@@ -13,7 +13,11 @@ from .visual import COLORS
 from .exceptions import IncompatibleAttribute
 from .component_link import (ComponentLink, CoordinateComponentLink,
                              BinaryComponentLink)
-from .subset import Subset, InequalitySubsetState, SubsetState
+from .subset import (Subset, InequalitySubsetState, SubsetState,
+                     RoiSubsetState, RangeSubsetState,
+                     CategoricalRoiSubsetState, AndState)
+from .roi import (PolygonalROI, CategoricalROI, RangeROI, RectangularROI,
+                    XRangeROI, YRangeROI)
 from .hub import Hub
 from .util import split_component_view, row_lookup
 from ..utils import unique, shape_to_string, view_shape, coerce_numeric, check_sorted
@@ -238,6 +242,55 @@ class Component(object):
     def jitter(self, method=None):
         raise NotImplementedError
 
+    def subset_from_roi(self, att, roi, other_comp=None, other_att=None, coord='x'):
+        """ Create a SubsetState object from an ROI.
+
+        This encapsulates the logic for creating subset states with
+        Components. See the documentation for CategoricalComponents for caveats
+        involved with mixed-type plots.
+
+        :param att: attribute name of this Component
+        :param roi: an ROI object
+        :param other_comp: The other Component for 2D ROIs
+        :param other_att: The attribute name of the other Component
+        :param coord: The orientation of this Component
+        :param is_nested: True if this was passed from another Component.
+        :return: A SubsetState (or subclass) object
+        """
+
+        assert coord in set('xy')
+
+        if hasattr(roi, 'min') and hasattr(roi, 'ori'):
+            # RangeROI and its subclasses
+            assert roi.ori in set('xy')
+            lo, hi = roi.range()
+            if roi.ori == coord:
+                subset_state = RangeSubsetState(lo, hi, att)
+            elif (roi.ori != coord) and isinstance(other_comp, CategoricalComponent):
+                other_coord = 'y' if coord == 'x' else 'x'
+                return other_comp.subset_from_roi(other_att, roi,
+                                                  other_comp=self,
+                                                  other_att=att,
+                                                  coord=other_coord)
+            elif roi.ori != coord:
+                subset_state = RangeSubsetState(lo, hi, other_att)
+            else:
+                raise AssertionError
+        else:
+            if isinstance(other_comp, CategoricalComponent):
+                return other_comp.subset_from_roi(other_att, roi,
+                                                  other_comp=self,
+                                                  other_att=att,
+                                                  is_nested=True)
+
+            subset_state = RoiSubsetState()
+            subset_state.xatt = att
+            subset_state.yatt = other_att
+            x, y = roi.to_polygon()
+            subset_state.roi = PolygonalROI(x, y)
+
+        return subset_state
+
     def to_series(self, **kwargs):
         """ Convert into a pandas.Series object.
 
@@ -437,6 +490,52 @@ class CategoricalComponent(Component):
         self._data = row_lookup(self._categorical_data, self._categories)
         self.jitter(method=self._jitter_method)
         self._data.setflags(write=False)
+
+    def subset_from_roi(self, att, roi, other_comp=None,
+                        other_att=None, coord='x',
+                        is_nested=False):
+        """ Create a SubsetState object from an ROI.
+
+        This encapsulates the logic for creating subset states with
+        CategoricalComponents. There is an important caveat, only RangeROIs
+        and RectangularROIs make sense in mixed type plots. As such, polygons
+        are converted to their outer-most edges in this case.
+
+        :param att: attribute name of this Component
+        :param roi: an ROI object
+        :param other_comp: The other Component for 2D ROIs
+        :param other_att: The attribute name of the other Component
+        :param coord: The orientation of this Component
+        :param is_nested: True if this was passed from another Component.
+        :return: A SubsetState (or subclass) object
+        """
+
+        assert coord in set('xy')
+
+        if hasattr(roi, 'min') and hasattr(roi, 'ori'):
+            # RangeRoi and its subclasses
+            assert roi.ori in set('xy')
+            if roi.ori == coord:
+                return CategoricalRoiSubsetState.from_range(self, att, roi.min, roi.max)
+            elif roi.ori != coord:
+                other_coord = 'y' if coord == 'x' else 'x'
+                return other_comp.subset_from_roi(other_att, roi,
+                                                  other_comp=self,
+                                                  other_att=att,
+                                                  coord=other_coord)
+            else:
+                raise AssertionError
+        elif isinstance(roi, CategoricalROI):
+            return CategoricalRoiSubsetState(roi=roi, att=att)
+        else:
+            x, y = roi.to_polygon()
+            if is_nested:
+                x, y = y, x
+
+            cat_subset = CategoricalRoiSubsetState.from_range(self, att,
+                                                              min(x), max(x))
+            cont_subset = RangeSubsetState(min(y), max(y), other_att)
+            return AndState(cat_subset, cont_subset)
 
     def jitter(self, method=None):
         """